AI-server/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
sniffio==1.3.1
python-dotenv==1.2.2
langchain-community==0.4.1
langchain-chroma==0.2.6
langchain-core==1.2.23
langchain-text-splitters==1.0.0
pypdf==6.9.2
docx2txt==0.9
zhipuai==2.1.5.20250825